+++ src/glproject/RealToVector3dFunction.java
@@ -2,5 +2,5 @@ package glproject;
 
 
 public interface RealToVector3dFunction {
-    public Vector3d eval(float t);
+    public Vector3f eval(float t);
 }
